requirements.txt
setuptools>=68.0.0
wheel>=0.41.0
Flask>=2.3.3
numpy>=1.25.0
opencv-python>=4.8.0
Pillow>=10.0.0
werkzeug>=2.3.7
